Remove legacy AC_CHECK_TYPE calls for uint and ulong
The AC_CHECK_TYPE was refactored in more recent versions of Autoconf
and the call with two arguments is obsolete and not recommended anymore.

This patch also refactors some leftovers of using ulong and uint which
are not standard nor common usages of types in C.

The ulong can be used as zend_ulong and uint usage is actually
`unsigned int`.

The usage of HAVE_ULONG removed since it is not used in current code
base.

Legacy edgecase for some legacy HPUX systems removed:
- sys/stream.h header is not checked and the HAVE_SYS_STREAM_H is
  not defined with current build system.

- flags are unsigned int
- max_allowed_packet changed to unsigned int

+++ ext/mysqlnd/mysqlnd_structs.h
@@ -206,7 +206,7 @@ typedef struct st_mysqlnd_infile
 
 typedef struct st_mysqlnd_session_options
 {
-	ulong		flags;
+	unsigned int		flags;
 
 	/* init commands - we need to send them to server directly after connect */
 	unsigned int	num_commands;
@@ -231,7 +231,7 @@ typedef struct st_mysqlnd_session_options
 
 	char 		*charset_name;
 	/* maximum allowed packet size for communication */
-	ulong		max_allowed_packet;
+	unsigned int		max_allowed_packet;
 
 #ifdef MYSQLND_STRING_TO_INT_CONVERSION
 	zend_bool	int_and_float_native;
